requirements.txt
ipykernel
langchain
chromadb
sentence-transformers
pypdf
langchain-community
pymupdf
huggingface_hub[hf_xet]
hf_xet
pandas
langchain
transformers
accelerate
bitsandbytes
langchain_google_genai
python-dotenv
cohere==5.15.0
langchain-huggingface
langchain-chroma
langchain-cohere
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart